% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_interactive_plot.R
\name{bb_interactive_plot}
\alias{bb_interactive_plot}
\title{Interactive logger data plot}
\usage{
bb_interactive_plot(
  d,
  dl,
  focal_column,
  jump_pattern,
  focal_flag_pattern,
  plot_label,
  focal_flag_label = plot_label,
  y_label,
  threshold_values,
  threshold_labels,
  threshold_colors,
  range_limit = c(NA, NA)
)
}
\arguments{
\item{d}{Data frame with logger data as produced by \code{\link[=qc_deployment]{qc_deployment()}}}

\item{dl}{Long version of the same data.}

\item{focal_column}{A single column from \code{d}}

\item{jump_pattern}{Regular expression to identify jump flags for the
focal columns.}

\item{focal_flag_pattern}{Regular expression to identify all flags for
the focal columns}

\item{plot_label}{Label for the plot.}

\item{focal_flag_label}{Label for the focal flag, " Flag" will be appended.}

\item{y_label}{Y axis label.}

\item{threshold_values}{One or more threshold values to be plotted as
horizontal lines.}

\item{threshold_labels}{The labels (text) to use for the threshold lines,
one per \code{threshold_value}.}

\item{threshold_colors}{The color to use for the threshold lines, can be
one color or a color for each \code{threshold_value}.}

\item{range_limit}{The plot range will never exceed this range - it can
be more restricted.}
}
\value{
A plotly plot object.
}
\description{
See the rmarkdown document
included with the package for example usage, found by running:
\code{system.file("rmd/QAQC_report.Rmd", package = "BuzzardsBay")}.
}
\details{
\code{bb_interactive_plot()} uses the \strong{plotly} package to make an interactive
plot of data from a Buzzard's Bay logger deployment.
}
\keyword{internal}
