% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_options.R
\name{bb_options}
\alias{bb_options}
\title{Set and retrieve BuzzardsBay package options}
\usage{
bb_options(...)
}
\arguments{
\item{...}{One of:
\enumerate{
\item one or more named arguments where the name is a
an option and the value its new setting e.g. \code{sal_max_jump = 0.75};
\item a single unnamed argument stating an option to retrieve e.g.
\code{"sal_max_jump"}.
\item No arguments, indicating that all options and their current settings
should be returned in a list; or
\item a single list argument with named items and their new values.
}}
}
\value{
If no arguments are used than all options will be returned as a list.
If there is a single, unnamed argument with a character value indicating an
option than the value of that option will be returned. Otherwise, the
arguments should indicate new option settings and nothing will be returned.
}
\description{
With no arguments all the Buzzards Bay parameters will be returned as a list.
Use a single character value to retrieve the value of a single option.
Use one or more named arguments to set options.
}
\details{
The options:
\describe{

\item{\code{do_lv_duration}, \code{do_lv_range}}{For Low Variation
in Dissolved Oxygen (\code{Dlv}) flag. \code{do_lv_range} is the maximum difference
between the maximum and the minimum values in a streak longer than
\code{do_lv_duration} minutes before the the low variation in dissolved oxygen
(\code{Dlv}) flag is set}
\item{\code{do_max_jump}}{The maximum difference between consecutive DO
readings before the jump in dissolved oxygen (\code{Dj}) flag is is set.}
\item{\code{do_streak_duration}, \code{do_streak_min}}{If DO remains below
\code{do_streak_min}
for more than \code{do_streak_duration} than the Dissolved Oxygen low streak
(\code{Dls}) flag is set.}
\item{\code{logger_error_values}}{One or more values that indicate a logger
error.  If setting multiple values in a YAML file use indented lines
with a dash for each value:

\if{html}{\out{<div class="sourceCode">}}\preformatted{logger_error_values:
  - -888.88
  - 9999
}\if{html}{\out{</div>}}

Temperature (from both loggers), High Range, and Raw DO are all checked
for this value. Flags: \code{TDe}, \code{TCe}, \code{He}, and \code{Re}.}
\item{\code{max_hr}, \code{min_hr}}{Thresholds for the high high range (\code{Hh}) and low
high range (\code{Hl}) flags.}
\item{\code{max_raw_do}}{Threshold for the high raw DO (\code{Rh}) flag.}
\item{\code{max_temp}, \code{min_temp}}{Thresholds for the high temperature
(\code{TDh}, \code{TCh}) and low temperature (\code{TDl}, \code{TCl})}
\item{\code{plot_min_do}, \code{plot_max_do}, \code{plot_min_sal}, \code{plot_max_sal},
\code{plot_min_temp}, \code{plot_max_temp}}{ These constrain the Y range in
the QC Report plots when plotting Dissolved Oxygen (\code{DO}),
Salinity (\code{sal}), and temperature (\code{temp})}
\item{\code{sal_lv_duration}, \code{sal_lv_range}}{If the difference between the
maximum and minimum salinity remains below \code{sal_lc_range} for more than
\code{sal_lv_duration} minutes than the low variation in salinity (\code{Slv}) flag
is set.}
\item{\code{sal_max_jump}}{The maximum difference between successive salinity
records before the salinity jump (\code{Sj}) flag is set for both involved
records.}
\item{\code{min_depth}, \code{max_depth}}{ The minimum and maximum water depth. If
the \code{Depth} column is present and falls outside of this range either the
\code{Wl} (water depth low) or \code{Wh} (water depth high) flags will be thrown.
Additionally \code{7} will be written to \code{Depth_QC} and either \code{9999}  or
\code{7} will be written to \code{Gen_QC} depending on if it's high or low
respectively}

}
}
\examples{
 o <- bb_options()
 cat(yaml::as.yaml(o))
 bb_options(sal_max_jump = 0.5)
 bb_options("sal_max_jump")
 bb_options(o)  # Reset original options
 bb_options("sal_max_jump")

}
