% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plot.R
\name{bb_plot}
\alias{bb_plot}
\title{Make static plot of logger data}
\usage{
bb_plot(
  d,
  dl,
  focal_columns,
  jump_pattern,
  focal_flag_pattern,
  plot_label,
  focal_flag_label = plot_label,
  y_label,
  threshold_values,
  threshold_labels,
  threshold_colors,
  lat = NULL,
  lon = NULL
)
}
\arguments{
\item{d}{Data frame with logger data as produced by \code{\link[=qc_deployment]{qc_deployment()}}}

\item{dl}{Long version of the same data.}

\item{focal_columns}{Names of columns in \code{d} that should be plotted,
max of 2 and they should have comparable scales.}

\item{jump_pattern}{Regular expression to identify jump flags for the
focal columns.}

\item{focal_flag_pattern}{Regular expression to identify all flags for
the focal columns}

\item{plot_label}{Label for the plot.}

\item{focal_flag_label}{Label for the focal flag, " Flag" will be appended.}

\item{y_label}{Y axis label.}

\item{threshold_values}{One or more threshold values to be plotted as
horizontal lines.}

\item{threshold_labels}{The labels (text) to use for the threshold lines,
one per \code{threshold_value}.}

\item{threshold_colors}{The color to use for the threshold lines, can be
one color or a color for each \code{threshold_value}.}

\item{lat, lon}{The latitude and longitude of the site, used to calculate
sunrise and sunset times.}
}
\value{
A ggplot plot object. Use \code{print()} to turn it into a plot/
}
\description{
\code{bb_plot()} uses \strong{gplot2} to generate plots of logger data. It is
primarily intended for use in reports.
See the rmarkdown document
included with the package for example usage, found by running:
\code{system.file("rmd/QAQC_report.Rmd", package = "BuzzardsBay")}.
}
