% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_placement.R
\name{check_placement}
\alias{check_placement}
\title{Check SN, site, deployment date, and logger type against placements table}
\usage{
check_placement(sn, type, placements, deployment_date, site)
}
\arguments{
\item{sn}{The logger serial number}

\item{type}{The logger type: \code{"DO"}, or \code{"Cond"}}

\item{placements}{The placements table}

\item{deployment_date}{The deployment date (end of deployment window). A
deployment is a single dip in the water ending with a data download.}

\item{site}{The site where the logger was deployed}
}
\value{
An informative error is thrown if something is wrong.
Nothing is returned.
}
\description{
This internal function verifies that the logger with the given SN was
placed at the expected, site, at the deployment date and that it is of the
expected type.
}
\keyword{internal}
