% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_site.R
\name{check_site}
\alias{check_site}
\title{Check to be sure that result files from \code{stitch_site} are up to date}
\usage{
check_site(site_dir, check_report = TRUE)
}
\arguments{
\item{site_dir}{Full path to site data (i.e., \verb{<base>/<year>/<site>}). The path must include
QCed results and result files from \code{stitch_site}.}

\item{check_report}{If TRUE, check to see if the report is up to date. Don't check when calling from
report_site, of course.}
}
\value{
Silently returns TRUE if validation was successful.
}
\description{
Check that source files for a site and year haven't changed since \code{stitch_site} was used
to create result files. Uses \code{hash.txt}, created by \code{check_site}. Separately reports on
missing hash file, missing source or result files, and changed source or result files.
}
\details{
Note that check_site does not throw errors--it simply tells you what's wrong. A calling
function (currently, just report_site) can use the silent return value to throw appropriate errors.
}
