% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_flags.R
\name{combine_flags}
\alias{combine_flags}
\title{Combine the flags from individual "_Flags" columns into
a single flags vector.}
\usage{
combine_flags(d)
}
\arguments{
\item{d}{A data frame.  Columns with "flag" in the name (ignoring case)
with the exception of a column called "Flags" (case specific) will
be concatinated into a return vector.}
}
\value{
a vector of the combined flags
}
\description{
This is a internal function called several times from \code{qc_deployment()}
}
\keyword{internal}
