% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_gmt.R
\name{convert_to_gmt}
\alias{convert_to_gmt}
\title{Convert date-time text to date time object in UTC}
\usage{
convert_to_gmt(dt, tz)
}
\arguments{
\item{dt}{One or more date time strings e.g. "2024-05-28 3:30:00"}

\item{tz}{A timezone like thing expressing the offset from GMT e.g.
"GMT-4:00"  This is the format used by hoboware and saved into the metadata
object and file.}
}
\value{
A date time object in UTC it will be offset from
the local "timezone".
}
\description{
\code{convert_to_gmt()}  Takes date time text in \if{html}{\out{<year>}}-\if{html}{\out{<month>}}-\if{html}{\out{<day>}} h:m:s and
for a "timezone" expressed as an offset from UTC/GMT and converts it into
a UTC time.
}
\details{
HOBOware is configured to log the time based on an offset from GMT and
it stores that offset.  This takes the time with the offset and removes the
offset to get it back to GMT and converts to a date time object with the
GMT timezone.
}
\examples{
convert_to_utc("2024-05-28 3:30:00", "GMT-4:00")

}
