% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_utc.R
\name{convert_to_utc}
\alias{convert_to_utc}
\title{Convert the local date-time text or date object to the equivalent date time
in UTC with a correctly defined (UTC) timezone.}
\usage{
convert_to_utc(dt, tz)
}
\arguments{
\item{dt}{One or more date time strings e.g. "2024-05-28 3:30:00" or
a date object made from such a string without specifying the timezone.  It
should be in the "timezone" indicated by \code{tz}.}

\item{tz}{A "timezone" string expressing the offset from GMT e.g.
"GMT-4:00"  This is the format used by HOBOware and saved into the metadata
object and file, but is not a standard timezone.}
}
\value{
A date time object in UTC that corresponds to the input time
(in \code{tz}).
}
\description{
\code{convert_to_utc()}  Takes date time text in \if{html}{\out{<year>}}-\if{html}{\out{<month>}}-\if{html}{\out{<day>}} h:m:s and
a separate "timezone" expressed as an offset from UTC/GMT and converts the
date times into UTC.
}
\details{
HOBOware is configured to log the time based on an offset from GMT.
The BuzzardsBay package stores that "timezone", but the offset isn't a valid
timezone. This gets around that by adding the offset hours and then defining
the timezone as "UTC".

For plotting I just use the local time so this isn't necessary, but the UTC
times are necessary for looking up the correct sunrise and sunset times.
}
