% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter_file.R
\name{create_parameter_file}
\alias{create_parameter_file}
\title{Create a BuzzardsBay parameter file}
\usage{
create_parameter_file(dir, overwrite = FALSE)
}
\arguments{
\item{dir}{The path to a directory to write the file to generally either
the base directory for the data or a site directory.}

\item{overwrite}{if \code{TRUE} overwrite any existing file.}
}
\value{
Nothing is returned. A file is written to disk
}
\description{
The parameter file is used to set parameters that affect how
flags are calculated. See \code{\link[=bb_options]{bb_options()}} for parameter descriptions.
}
