% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_path_items.R
\name{cut_path_items}
\alias{cut_path_items}
\title{cut items off the end of a path}
\usage{
cut_path_items(x, ncut)
}
\arguments{
\item{x}{A path}

\item{ncut}{The number of items to cut}
}
\value{
A shortened path
}
\description{
cut items off the end of a path
}
\examples{
\dontrun{
  deployment_dir <- "C:/temp/bb_demo//BB_Data/2023/RB1/2023-06-09"
  base_dir <- cut_path_items(deployment_dir, 3)
}
}
\keyword{internal}
