% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_stats.R
\name{daily_stats}
\alias{daily_stats}
\title{Produce daily stats for a site}
\usage{
daily_stats(site, year, core)
}
\arguments{
\item{site}{Three-letter site abbreviation}

\item{year}{Year of deployment}

\item{core}{Core data frame, produced by \code{stitch_site}}
}
\value{
Data frame with one row for each day of depolyment, and columns with a number of statistics
}
\description{
Produces a data frame of daily stats.
}
