% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_baywatchers.R
\name{extract_baywatchers}
\alias{extract_baywatchers}
\title{Extract relevant Baywatchers data from ginormous Excel file}
\usage{
extract_baywatchers(
  base_dir,
  year,
  file = "bbcdataCURRENT.xlsx",
  header_line = 3
)
}
\arguments{
\item{base_dir}{Path to BB_Data base directory}

\item{year}{Year to extract data for}

\item{file}{File name of Baywatchers Excel file. This file must be in the base directory.}

\item{header_line}{Line in the Excel with headers}
}
\description{
Pulls the relevant site, date, and DO values from the Baywatchers source file for a
given year. Rows with \code{DO_QC = 9} will be dropped. Rows with missing times (column \code{TIME})
will also be dropped. The results will be written to \verb{<year>/baywatchers.csv}, along with
a hash file, \code{bay_hash.txt}.
}
\details{
Run this before \code{report_site} if you want to include the two Baywatchers plots (run with
\code{baywatchers = TRUE}). It only needs to be run once for each year (or if the Baywatchers
data are updated). \code{check_site} will check to make sure \code{baywatchers.csv} is up to date
with respect to the source Excel file.
}
