% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_csv_date.R
\name{format_csv_date}
\alias{format_csv_date}
\title{Format CSV date that might have been corrupted by excel}
\usage{
format_csv_date(x)
}
\arguments{
\item{x}{Dates as character month/day/year or year-month-day can be
mixed in one vector.}
}
\value{
Dates as year-month-day (character)
}
\description{
Excel with US localization will automatically convert dates from
year-month-day to month/day/year.  Thus whenever a date is read from a
csv that might have been edited with Excel I'm going to check for slashes
and assume month/day/year if they are present.
}
\examples{
# example code
\dontrun{
x <- c(NA, "2024-05-15", "2024-05-16", "5/16/2024")
}
}
\keyword{internal}
