% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_date_time.R
\name{format_date_time}
\alias{format_date_time}
\title{Internal function to reformat the date time date from Details.txt}
\usage{
format_date_time(x)
}
\arguments{
\item{x}{Date time in a  month/day/year h:m:s GMT-offset format

Note this returns a string not a date time object, and the GMT offset is
dropped. If this string is processed into a date time object the timezone
will likely be wrong if not explicitly set.}
}
\value{
String with year-month-day h:m:s format.
}
\description{
Internal function to reformat the date time date from Details.txt
}
\examples{
\dontrun{
x <- "06/10/19 08:05:37  GMT-04:00"
format_date_time(x)
}
}
