% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_hoboware_details.R
\name{get_cond_details}
\alias{get_cond_details}
\title{Retrieve information from a conductivity calibration details file}
\usage{
get_cond_details(path)
}
\arguments{
\item{path}{The path to a details file created by HOBOware during
conductivity calibration.}
}
\value{
A nested list containing of a small subset of the information in
\code{file}.
}
\description{
\code{get_cond_details()} extracts a specific subset of the information stored
in a details file generated by HOBOware during conductivity/salinity
calibration.
}
\examples{
 file <- system.file(
   "extdata/2023/RB1/2023-06-09/Calibrated/Cond_RB1_2023-06-09_Details.txt",
                     package = "BuzzardsBay")
 do_cal  <- get_cond_details(file)
 cat(yaml::as.yaml(do_cal))

}
\seealso{
\itemize{
\item \code{\link[=get_do_details]{get_do_details()}} for a similar function to process
a HOBOware details file generated while calibrating dissolved oxygen.
\item \code{\link[=parse_hoboware_details]{parse_hoboware_details()}} for a function to read and format all
of the information in a HOBOware details file.
}
}
