% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_hoboware_details.R
\name{get_do_details}
\alias{get_do_details}
\title{Retrieve information stored a Dissolved Oxygen calibration details file}
\usage{
get_do_details(path)
}
\arguments{
\item{path}{The path to a details file created by HOBOware during
calibration.}
}
\value{
A nested list containing of a small subset of the information in
\code{file}.
}
\description{
\code{get_do_details()} extracts a specific subset of the information stored in a
details file generated by HOBOware during DO calibration.
}
\examples{
 file <- system.file(
   "extdata/2023/RB1/2023-06-09/Calibrated/DO_RB1_2023-06-09_Details.txt",
                     package = "BuzzardsBay")
 do_cal  <- get_do_details(file)
 cat(yaml::as.yaml(do_cal))

}
\seealso{
\itemize{
\item \code{\link[=get_cond_details]{get_cond_details()}} for a similar function to process
a HOBOware details file generated while calibrating conductance/salinity.
\item \code{\link[=parse_hoboware_details]{parse_hoboware_details()}} for a function to read and format all
of the information in a HOBOware details file.
}
}
