% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expected_columns.R
\name{get_expected_columns}
\alias{get_expected_columns}
\title{Look up expected column names and order}
\usage{
get_expected_columns(type, existing = character(0))
}
\arguments{
\item{type}{What type (point along the data pipeline) are the names for,
currenttly one of
\itemize{
\item \code{calibration}  output from \code{import_calibrated_data()}
\item \code{intermediate}  intermediate in QC proces this includes the individual
flag columns.
\item \code{qc_final}  output from \code{qc_deployment()}
}}

\item{existing}{A vector of existing column names. This is used to determine
which of the optional column names are present}
}
\value{
A list of the expected names (including only the optional names
that are present) in canonical order.
}
\description{
\code{get_expected_columns()} looks up the column names and order expected at
different points in the data pipeline.  The columns them selves are
set in the \code{expected_columns}  list object.
}
\details{
At each step in the process there are both required and optional names
this function handles the fuzzyness between those two - returning a final
set of column names, in canonical order.

This function uses lists of column names defined separately in
\code{R/expected_column_names.R} edit that list to change the output.
}
\keyword{internal}
