% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_hashes.R
\name{get_file_hashes}
\alias{get_file_hashes}
\title{Get md5 hashes for a vector of files}
\usage{
get_file_hashes(files)
}
\arguments{
\item{files}{A list of full file specifications}
}
\value{
A matching vector of md5 hashes
}
\description{
Given one or more file paths and names, return md5 hashes for each. Returns
NA for files that don't exist.
}
\keyword{internal}
