% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_name.R
\name{get_site_name}
\alias{get_site_name}
\title{Get the full site name for a site}
\usage{
get_site_name(site_dir)
}
\arguments{
\item{site_dir}{Site directory}
}
\value{
Full name of the site
}
\description{
Pulls the full site name from the sites table.
}
\keyword{internal}
