% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tide_height.R
\name{get_tide_height}
\alias{get_tide_height}
\title{Get the tide height at a particular station corresponding to a series
of date, times. Note due to limitations in the \strong{rtide} package this
only works for about a third of the Buzzards Bay NOAA tide stations.}
\usage{
get_tide_height(dt, station)
}
\arguments{
\item{dt}{Series of date-time values as POSIXct. The timezone must be
properly set.}

\item{station}{A station name or station ID as it appears in
\link{tide_station_info}.}
}
\value{
A series of tide heights corresponding to the date-times in \code{dt}
Heights are in meters and, I think, are relative to
\href{https://tidesandcurrents.noaa.gov/datum_options.html#MLLW}{MLLW}
(Mean Lower Low  Water). Although, I could only find this stated about
individual stations and not for all stations in general.

See for example the datum information for
\href{https://tidesandcurrents.noaa.gov/datums.html?id=8448248}{PENIKESE ISLAND MA},
which states "NOTICE: All data values are relative to the MLLW.".
}
\description{
\code{get_tide_height()} is a wrapper to \code{\link[rtide:tide_height]{rtide::tide_height()}} that works
with a date time vector and also checks the station against
\code{tide_station_info}.
}
