% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tide_rider.R
\name{import_tide_rider}
\alias{import_tide_rider}
\title{Import Tide Rider Data}
\usage{
import_tide_rider(paths)
}
\arguments{
\item{paths}{paths list as returned from \code{\link[=lookup_paths]{lookup_paths()}} the only
component used here is \verb{$deployment_cal_dir}.  It should contain a
csv file with \code{"TR"} bounded by \code{"_"} or  \code{"."}  e.g. \code{"_TR_"} or \code{"_TR."}.}
}
\value{
Tide Rider data with columns:
\item{Date_Time}{Date and time of each record as character in the
format year-month-day h:m:s (all numeric). The file can use that format
or day-month-year h:m:s where month is an abbreviated month name e.g.
"12-Aug-2025 13:40:00".}
\item{Latitude, Longitude}{Location in spherical coordinates. Assumed to
be WGS84 (EPSG:4326)}
\item{Depth}{Logger depth in meters}
\item{TR_Flags}{Tide Rider Flags.}
}
\description{
This imports the location information for a Tide Rider from a CSV file.
See \code{\link[=import_calibrated_data]{import_calibrated_data()}} for more information.
}
\keyword{internal}
