% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_daylight.R
\name{is_daylight}
\alias{is_daylight}
\title{Determine if observation correspond to daylight hours or not}
\usage{
is_daylight(dt, lat, lon, tz)
}
\arguments{
\item{dt}{A vector of date times either as text or date time objects}

\item{lat, lon}{The latitude and longitude of the associated site.}

\item{tz}{The HOBOware style GMT offset used to define the "timezone" of the
dates in \code{dt}. Probably always "GMT-4:00" but use the value saved to the
metadata file.}
}
\value{
A date time object with the dates in \code{dt} converted to UTC and with
that timezone correctly set in the object.
}
\description{
Determine if observation correspond to daylight hours or not
}
