% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_devices.R
\name{lookup_devices}
\alias{lookup_devices}
\title{Lookup the DO and conductivity logger device models}
\usage{
lookup_devices(site, deployment_date, placements)
}
\arguments{
\item{site}{The site where the logger was deployed}

\item{deployment_date}{The deployment date (end of deployment window). A
deployment is a single dip in the water ending with a data download.}

\item{placements}{The placements table}
}
\value{
A list with items:
\item{do_model,do_sn}{ Dissolved oxygen sensor model and serial number}
\item{cond_model,cond_sn}{ Conductivity sensor model and serial number}
}
\description{
Given a deployment date and site lookup the sensors deployed there.
The assumption is that there is both a \code{"Do"} and
\code{"Cond"} device type associated with each placement and this function
will throw and error if that is violated for the specified deployment.
}
