% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_paths.R
\name{lookup_paths}
\alias{lookup_paths}
\title{Lookup paths for data and parameter files}
\usage{
lookup_paths(
  base_dir = NULL,
  year = NULL,
  site = NULL,
  deployment_date = NULL,
  deployment_dir = NULL
)
}
\arguments{
\item{base_dir}{The base directory for all the data.}

\item{year}{The four digit year for which paths are to be looked up}

\item{site}{The site name for which the paths are to be looked up}

\item{deployment_date}{The deployment date in the form \code{yyyy-mm-dd}, where
\code{yyyy} is a four digit year, \code{mm} is a two digit month and \code{dd} is a two
digit day.  Single digit components should be padded e.g. \code{01} for January.}

\item{deployment_dir}{(optional) The full path to a deployment's directory.
If \code{deployment_dir} is used, all other arguments are ignored and instead
their values are inferred from this path.}
}
\value{
A list of paths and a few related non-path variables.  The items
returned are all always the same and in the same order but with some inputs
the deployment specific items will not be defined so will have \code{NA} values.
All list items contain a single string.
\item{base_dir}{The path to the base directory for the data.}
\item{year_dir}{The path to a specific year's directory}
\item{site_dir}{The path to a specific site's directory}
\item{md_dir}{The path to a metadata  / parameter directory. Containing
additional information about sites and placements for the given year}
\item{deployment_dir}{The path to the deployment directory where all the
data (raw, calibrated, and QC'd) for a deployment is stored.  A deployment
is the period a device is continuously in the water preceding data
download from the device, typically about a week for this project.}
\item{deployment_cal_dir}{The deployment calibration directory path}
\item{deployment_auto_qc}{The deployment Auto QC file path this file is
created by \code{\link[=qc_deployment]{qc_deployment()}} as a permanent record of the output of the
automated QC process. It should not be modified.}
\item{deployment_prelim_qc}{This is the preliminary QC file. It is created
by \code{\link[=qc_deployment]{qc_deployment()}} and is intended to be edited by people during the AC
process who will, at completion of that process delete the \code{"preliminary_"}
from the beginning of the name.}
\item{deployment_final_qc}{This is the final QC file created from
\code{deployment_prelim_qc} by people performing the QC.}
\item{deployment_metadata}{This file contains metadata from the HOBOware
Details files and generated by \code{\link[=qc_deployment]{qc_deployment()}}.}
\item{deployment_report}{An html report with plots to inform the QC process.}
\item{preceding_auto_qc}{The path to the auto QC file for the preceding
deployment if it exists, NA otherwise.}
\item{sites}{CSV file with sites and associated information. The sites file
is  maintained by the users.}
\item{placements}{CSV tracking placements.  The placement is the period a
specific device is at a site. The placements file is maintained by
the users.}
\item{import_types}{File connecting device models (in placements file) to
the import type, an integer, used to import the calibrated data derived
from those device models each import type must be supported with a
function named \verb{import_calibrated_data_\{type\}}. where type the integer.
See for example \code{import_calibrated_data_1} which handles the CSV output
from HOBOware.
The public function \code{\link[=import_calibrated_data]{import_calibrated_data()}} handles dispatch
to the numbered functions and does some additional formatting and checking
of the output.}
\item{global_parameters}{This YAML (\code{.yml}) file sets global
parameters for the BuzzardsBay package, most of which affect how flags
are calculated.  See \code{update_bb_parameters}}
\item{site_parameters}{This YAML takes precedence over
\code{global_parameters} and allows adjusting them to a particular site}
\item{year}{NOT a path This is the four digit year in character format}
\item{deployment_date}{NOT a path. This is the deployment date as a
character in \code{yyyy-mm-dd} format.  Numbers are padded e.g. \code{01} for
January.}
\item{site}{NOT a path. This is the site name. It is case sensitive and
should appear exactly as it does in the sites table (\code{sites}).}
}
\description{
This includes all the paths that relate to a deployment, site, year, or
the entire data base.
}
\details{
The easiest way to call it is to specify the deployment directory
(\code{deployment_dir}) in which case all other arguments are inferred from the
path. Otherwise specify at a minimum \code{base_dir}, \code{year}, and \code{site}; and,
optionally \code{deployment_date} and all paths that can be specified will be.
}
