% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_site_paths.R
\name{lookup_site_paths}
\alias{lookup_site_paths}
\title{Look up paths for a specified site}
\usage{
lookup_site_paths(site_dir)
}
\arguments{
\item{site_dir}{Full path to site data. The path must include QCed results.}
}
\value{
A named list consisting of:
\item{sites}{The path to the site info table}
\item{deployments}{A data frame with a row for each deployment, and columns \code{year}, \code{date}, \code{QCpath}, \code{md_path}, and \code{hash},
containing the year and full date for each deployment, the path to the final QCed data, the path to the deployment metadata
file, and the md5 hash of the QCed data file.}
}
\description{
Find and return paths to all depolyment directories, metadata files, and final QC files for selected site.
}
