% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_deployment_report.R
\name{make_deployment_report}
\alias{make_deployment_report}
\title{Make QAQC report for a deployment}
\usage{
make_deployment_report(dir, quiet = FALSE)
}
\arguments{
\item{dir}{A deployment directory}

\item{quiet}{Set to \code{TRUE} for less chatter. Passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}
}
\value{
Invisibly returns the path the the QC report
}
\description{
\code{make_deployment_report()} Will make an html report to use while conducting
quality assurance and quality control (QAQC) on a deployment.  It
uses the "Auto_QC" CSV file generated by "qc_deployment()".
}
\details{
If an output ".html" file exists it will be rewritten.
}
\examples{
\dontrun{

example_paths <- setup_example_dir()
dir <- example_paths$deployment
qc_deployment(dir, report = FALSE)
report_path <-  make_deployment_report(dir)
}
}
