% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_facet_plot.R
\name{make_facet_plot}
\alias{make_facet_plot}
\title{Produce a paired time series plot of DO and Temperature for the final report}
\usage{
make_facet_plot(daily)
}
\arguments{
\item{daily}{daily data frame, from \code{report_site}}
}
\value{
A ggplot2 object
}
\description{
Produce a paired time series plot of DO and Temperature for the final report.
Unlike \code{make_plot}, this function is specific for a single plot.
}
\keyword{internal}
