% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot.R
\name{make_plot}
\alias{make_plot}
\title{Produce a time series plot for the final report}
\usage{
make_plot(
  core = core,
  x = core$Date_Time,
  y,
  x_lab,
  y_lab,
  hline = 0,
  diag = FALSE,
  point_size = 0.25,
  linetype = 0,
  vline = NULL
)
}
\arguments{
\item{core}{Core data frame, produced by \code{stitch_site}}

\item{x}{X variable, typically Date_Time}

\item{y}{Y variable}

\item{x_lab}{Label for x-axis}

\item{y_lab}{Label for y-axis}

\item{hline}{Optional horizontal line at specified value on the y-axis}

\item{diag}{If TRUE, set x and y axes to the same range and draw a diagonal
with a slope of 1}

\item{point_size}{Optional size of points in plot}

\item{linetype}{Optional line type (default 0, is no line)}
}
\value{
A ggplot2 object
}
\description{
Produce a time series plot for the final report. Values are plotted as
small points, with specified data, x and y columns, and axis labels. A
horizontal line can optionally be plotted.
}
\keyword{internal}
