% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_2var.R
\name{make_plot_2var}
\alias{make_plot_2var}
\title{Produce a time series plot with two datasets for the final report}
\usage{
make_plot_2var(data1, data2, x1, y1, x2, y2, x_lab, y_lab)
}
\arguments{
\item{data1}{Second dataset}

\item{x1}{Second X variable}

\item{y1}{Second Y variable}

\item{x_lab}{Label for x-axis}

\item{y_lab}{Label for y-axis}
}
\value{
A ggplot2 object
}
\description{
Produce a time series plot for the final report that includes 2 sets of variables. Values
in first set are plotted as small black points, and those in second set as larger red points,
plotted on top of the first.
}
\keyword{internal}
