% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Produce a time series plot for the final report}
\usage{
plot_ts(
  core = core,
  x = core$Date_Time,
  y,
  x_lab,
  y_lab,
  hline = 0,
  point_size = 0.25
)
}
\arguments{
\item{core}{Core data frame, produced by \code{stitch_site}}

\item{x}{X variable, typically Date_Time}

\item{y}{Y variable}

\item{x_lab}{Label for x-axis}

\item{y_lab}{Label for y-axis}

\item{hline}{Optional horizontal line at specified value on the y-axis}

\item{point_size}{Optional size of points in plot}
}
\value{
ggplot2 object
}
\description{
Produce a time series plot for the final report. Values are plotted as
small points, with specified data, x and y columns, and axis labels. A
horizontal line can optionally be plotted.
}
\keyword{internal}
