% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_deployment.R
\name{qc_deployment}
\alias{qc_deployment}
\title{Run automatic quality control on a deployment}
\usage{
qc_deployment(dir, report = TRUE)
}
\arguments{
\item{dir}{The path to a deployment directory to run QC on.  It should
have a \verb{Calibration/} directory with calibrated DO and conductivity data
associated Details files generated by HOBOware.}

\item{report}{If \code{TRUE} (the default) an html report will be generated
with plots to assist the QAQC process.  Set to \code{FALSE} to skip generating
the report. See \code{\link[=make_deployment_report]{make_deployment_report()}} to explicitly make the report
as a separate step.}
}
\value{
A list with items:
\item{d}{The final tabular data as written to disk.}
\item{md}{A nested list of the metadata.}
}
\description{
\code{qc_deployment()} will:
\enumerate{
\item Read the calibrated dissolved oxygen and conductivity files created by
HOBOware.
\item Read and parse the associated Details files into nested lists.
\item Reformat and merge the data files.
\item Add automatic flags, and create columns for manual QC
\item Write the data and metadata.
\item If \code{report = TRUE} it will call \code{\link[=make_deployment_report]{make_deployment_report()}} to make
and interactive HTML report with plots for QAQC.
}
}
\details{
\code{qc_deployment()} assumes the files are arranged in a specific
\href{https://docs.google.com/document/d/1kJttcEXzpNNknGwjkVwdYHw9LzZyjJ-FaX0CrU7H7NU/edit#heading=h.s6vs4d7vj0cn}{file structure}.
}
\note{
\code{qc_deployment()} calls \code{\link[=import_calibrated_data]{import_calibrated_data()}} which currently supports
three completely different import functions.  See \code{\link[=import_calibrated_data]{import_calibrated_data()}}
for documentation on the data formats they each expect.

In all cases input files are looked for within a \code{Calibration} sub directory
to the deployment directory (\code{dir}).
}
\examples{
\dontrun{
  paths <- setup_example_dir()
  a <- qc_deployment(paths$deployment)

}

}
