% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_format_placements.R
\name{read_and_format_placements}
\alias{read_and_format_placements}
\title{Read and format placements}
\usage{
read_and_format_placements(placement_path)
}
\arguments{
\item{placement_path}{path to placements file}
}
\value{
placement data_frame
In addition to reading the file the following changes are made:
\itemize{
\item column names headers are forced to lower case,
\item dates are formatted as dates while handling both the \code{yyyy-mm-dd} format
and the format that excel converts them to.
\item \code{type} is forced to lower case.
}
}
\description{
Read and format placements
}
