% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mx801_data.R
\name{read_mx801_data}
\alias{read_mx801_data}
\title{Import data from MX801 logger}
\usage{
read_mx801_data(file)
}
\arguments{
\item{file}{Path to a calibrated Excel (\code{.xlsx}) file from an MX801 logger
Column descriptions in
\href{https://docs.google.com/spreadsheets/d/1bYxi0nbgDaUsKEyLoj_ry-1Zc-A4MNDfS_J_zH7HiJI}{google doc}}
}
\value{
A data frame with standard data columns
}
\description{
Import data from MX801 logger
}
\keyword{internal}
