% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na_runs.R
\name{replace_na_runs}
\alias{replace_na_runs}
\title{Replace short runs of NA in logical vector}
\usage{
replace_na_runs(x, max_run = 6, boundary = TRUE)
}
\arguments{
\item{x}{A logical vector that may have NAs}

\item{max_run}{The maximum run length to replace}

\item{boundary}{How to treat values beyond the edge of the data}
}
\value{
A vector corresponding to \code{x} where qualifying runs have been replaced with TRUE
}
\description{
Any runs of NA not longer than \code{max_run} that are surrounded by TRUE are
replaced with TRUE. All other NAs are replaced with FALSE.
}
\details{
Use \code{boundary} to specify behavior at data boundaries. For instance, if bounadary = TRUE,
then c(NA, TRUE) would result in c(TRUE, TRUE); otherwise the result would be c(FALSE, TRUE).
}
