% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{\enumerate{
\item daily_stats.csv - a file with a row for each day of the season with several summary statistics.
}}
\usage{
report(site_dir)
}
\description{
\enumerate{
\item daily_stats.csv - a file with a row for each day of the season with several summary statistics.
}
}
