% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle2.R
\name{rle2}
\alias{rle2}
\title{A drop-in replacement for \code{base::rle()} that treats all NAs as identical}
\usage{
rle2(x)
}
\arguments{
\item{x}{an atomic vector}
}
\value{
An object of class \code{rle}
}
\description{
A drop-in replacement for \code{base::rle()} that treats all NAs as identical
}
\keyword{internal}
