% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_plots.R
\name{seasonal_plots}
\alias{seasonal_plots}
\title{Produce seasonal plots for a site}
\usage{
seasonal_plots(core)
}
\arguments{
\item{core}{Core data frame, produced by \code{stitch_site}}
}
\value{
List of ggplot2 objects
}
\description{
Produces a list with a series of seasonal plots as ggplot2 objects.
}
\keyword{internal}
