% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_example_dir.R
\name{setup_example_dir}
\alias{setup_example_dir}
\title{Setup an example directory}
\usage{
setup_example_dir(parent_dir = NULL)
}
\arguments{
\item{parent_dir}{The directory in which to create the example files tree.
If left \code{NULL} it will be set to \code{\link[=tempdir]{tempdir()}}. Either way
an error will be thrown if this a \code{"BB_Data"} folder already exists
in \code{parent_dir}.}
}
\value{
A named list with paths to:
\item{base_dir}{The path to the newly created example
base directory (\code{"BB_Data"}).}
\item{deployment_dir}{The path to the example deployment folder within
\code{base_dir}.}
}
\description{
Setup an example directory for testing and demonstrating the \strong{BuzzardsBay}
package functions. It will be created within \code{parent_dir} called \code{"BB_Data"}
and will contain example calibrated data files for a single deployment as
well as some of project level metadata.  The files are bundled with the
package and can be found in the source within \verb{inst/extdata/}.
}
\examples{
\dontrun{
setup_example()
}
}
