% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppress_specific_warnings.R
\name{suppress_specific_warnings}
\alias{suppress_specific_warnings}
\title{Suppress some specific warnings}
\usage{
suppress_specific_warnings(x, patterns = NULL)
}
\arguments{
\item{x}{An expression.}

\item{patterns}{One or patterns to check warning messages against.}
}
\value{
Possibly output from \code{x}
}
\description{
\code{suppress_specific_warnings()} will suppress warnings that match regular
expression patterns that are either coded into the function or supplied via
the \code{patterns} argument, without suppressing warnings that don't match the
patterns.
}
\details{
It's initial (possibly only) use is to catch a warning thrown by ggplot due
to NA value created by using lead() when plotting time of day.

\code{suppress_specific_warnings()} is used in the rmarkdown report so needs
to be a public function but it is not intended for external use.
}
\examples{
suppress_specific_warnings(warning("this is a warning"), "this is")
suppress_specific_warnings(warning("this is a warning"))

}
\keyword{internal}
