% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide_station_info.R
\docType{data}
\name{tide_station_info}
\alias{tide_station_info}
\title{Locations of NOAA tide stations within Buzzards Bay}
\format{
\subsection{\code{tide_station_info}}{

A data frame with 13 rows and 5 columns:
\describe{
\item{name}{Station name. This can be used with \link[rtide:tide_height]{rtide::tide_height}}
\item{id}{Station ID}
\item{lat,lon}{Station latitude and longitude}
\item{predictions}{
The type of model used to make predictions. One of \code{"Harmonic"} or
`"Subordinate".

Descriptions from \href{https://tidesandcurrents.noaa.gov/PageHelp.html}{NOAA}:

\strong{Harmonic stations} are stations with tidal harmonic constants and tidal
datums. Tide predictions for harmonic stations are generated directly
from the harmonic constants. Harmonic stations have the greatest
capabilities within the NOAA Tide Predictions service for providing
predictions with different data intervals and relative to different tidal
datums. These capabilities are controlled within the interface.

\strong{Subordinate stations} are historic stations which do not have tidal
harmonic constants available. Tide predictions for subordinate stations
are generated by first generating high/low tide predictions for a
designated harmonic station, called the “reference station”; then time and
height adjustments are applied to correct the high/low predictions to the
tidal conditions at the subordinate station. Subordinate stations only
allow for the generation of high/low tide predictions, and heights will be
relative to Mean Lower Low Water (MLLW), the standard chart datum for the
U.S. coastline.}

}
}
}
\source{
<"https://tidesandcurrents.noaa.gov/tide_predictions.html?gid=1403">
}
\usage{
tide_station_info
}
\description{
Locations of NOAA tide stations within Buzzards Bay
}
\keyword{datasets}
