% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_bb_parameters.R
\name{update_bb_parameters}
\alias{update_bb_parameters}
\title{Update Buzzards Bay parameters}
\usage{
update_bb_parameters(paths)
}
\arguments{
\item{paths}{Likely the output of \code{\link[=lookup_paths]{lookup_paths()}} the only used items
are \verb{$global_parameters} and \verb{$site_parameters.}}
}
\value{
Nothing is returned, but package parameters are updated.
}
\description{
Internal function to update the BuzzardsBay package parameters.  It sets
parameters based on the default parameters for the package,
a required global parameter file  (\code{paths$global_parameters}) and an
optional site specific parameter file (\code{paths$site_parameters}).
}
\details{
The same parameter may be set in multiple locations in which case the
precedence in decreasing
order is site parameter file, global parameter file, and package default
value \code{default_bbp}.  The currently set parameters are stored in the \code{bbp}
environment within the package.

Note, not all parameters need to be set in all locations. It is recommended
to set all the parameters in the global file and then only ones that need
to be changed in the site specific file.

See \code{\link[=create_parameter_file]{create_parameter_file()}} to create a file with default values.
}
\keyword{internal}
