% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_calibrated_data.R
\name{import_calibrated_data}
\alias{import_calibrated_data}
\title{Import Calibrated Data}
\usage{
import_calibrated_data(paths, devices)
}
\arguments{
\item{paths}{Path list as returned from \verb{[lookup_paths()]}.  Only the
\code{import_types} component is used here.}

\item{devices}{Device information in a list as returned from
\code{\link[=lookup_devices]{lookup_devices()}}.}
}
\value{
A list with two items the first \code{d} is a data frame with the
calibrated data containing the following columns.
\item{Date_Time}{The date and time of the observation as a string in the
format \verb{yyyy-mm-dd hh:mm:ss}}
\item{Raw_DO}{Uncalibrated (measured) Dissolved Oxygen in mg/L }
\item{Temp_DOLog}{Temperature in Deg. Celsius as recorded by the DO logger}
\item{DO}{The calibrated (salinity adjusted) Dissolved Oxygen in mg/L}
\item{DO_Pct_Sat}{Dissolved oxygen percent saturation (calibrated)}
\item{Salinity_DOLog}{Salinity as recorded by the DO logger in ppt
(parts per thousand), this is equivalent to PSU (Practical Salinity Units) }
\item{High_Range}{High Range Conductivity in microsiemens per cm. Note newer
loggers may call this column "Electrical Conductivity" which is the same
thing, and even for those loggers it will be called High_Range throughout
the QC workflow.}
\item{Temp_CondLog}{Temperature in Deg. Celsius as recorded by the
conductivity logger}
\item{Spec_Cond}{The specific conductivity (new loggers) or specific
conductance (older loggers) in microsiemens per cm.}
\item{Salinity}{Salinity as recorded by the Conductivity logger in ppt
(parts per thousand), equivalent to PSU (Practical Salinity Units) which
is the units assigned by the newer loggers}
}
\description{
This function imports calibrated data from \code{paths$deployment_cal_dir}
it calls distinct functions depending on the devices associated with the
deployment based on the \code{import_type} associated with the devices in
the  \code{import_type.csv} file in the base directory for the Buzzards Bay
data.
}
\details{
Import types:
\enumerate{
\item The original input type that used calibrated data exported from
Hobo ware.  This expects to find two CSV files and two \code{detailes.txt}
files in the calibration directory. The Dissolved Oxygen files should have
\code{"Do_"} somewhere in the names while the Conductivity files should have
either \code{"Cond_"} or \verb{"Sal_} in the names.
\item This is the import type associated with the \code{MX801} logger. Note
there should still be two lines in \code{placements.csv} one each for
\code{"Do"} and \code{"Cond"} both of which should indicate "MX801".  This import
expects a single  excel file with combined data from both loggers and on
one sheet and the details information on another.  NOT IMPLEMENTED YET.
}
}
