% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longest_duration.R
\name{longest_duration}
\alias{longest_duration}
\title{Give duration of the longest run of low DO for each day}
\usage{
longest_duration(x)
}
\arguments{
\item{x}{A data frame with three columns:
\enumerate{
\item \code{Date} The date, in format 2024-05-30
\item \code{Date_Time} The date and time, in format 2024-05-30 16:20:00
\item \code{trigger} TRUE when DO is below the selected threshold (the name of this column varies)
}}
}
\value{
Length of the longest run of TRUEs in fractional hours
}
\description{
Finds the longest run of DO below a threshold for each day.
}
