% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longest_duration.R
\name{longest_duration}
\alias{longest_duration}
\title{Give the longest run of TRUEs}
\usage{
longest_duration(x)
}
\arguments{
\item{x}{Data frame with columns \code{Date_Time} and \code{trigger} (TRUE when the trigger condition is met)}
}
\value{
Length of the longest run of TRUEs in hh:mm
}
\description{
Used to find the longest run of DO below a threshold.
}
