% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_deployment.R
\name{qc_deployment}
\alias{qc_deployment}
\title{Run automatic quality control on a deployment}
\usage{
qc_deployment(dir, report = TRUE)
}
\arguments{
\item{dir}{The path to a deployment directory to run QC on.  It should
have a \verb{Calibration/} directory with calibrated DO and conductivity data
associated Details files generated by HOBOware.}

\item{report}{If \code{TRUE} (the default) an html report will be generated
with plots to assist the QAQC process.  Set to \code{FALSE} to skip generating
the report. See \code{\link[=make_deployment_report]{make_deployment_report()}} to explicitly make the report
as a separate step.}
}
\value{
A list with items:
\item{d}{The final tabular data as written to disk.}
\item{md}{A nested list of the metadata.}
}
\description{
\code{qc_deployment()} will:
\enumerate{
\item Read the calibrated dissolved oxygen and conductivity files created by
HOBOware.
\item Read and parse the associated Details files into nested lists.
\item Reformat and merge the datafiles.
\item Add automatic flags, and create columns for manual QC
\item Write the data and metadata.
}
}
\details{
\code{qc_deployment()} assumes the files are arranged in a specific
\href{https://docs.google.com/document/d/1kJttcEXzpNNknGwjkVwdYHw9LzZyjJ-FaX0CrU7H7NU/edit#heading=h.s6vs4d7vj0cn}{file structure}.
}
\note{
Input files are looked for within \code{file.path(dir, "Calibration")}
using a heuristic that assumes the detail files end in \code{"Details.txt"}; the
data files end in \code{".csv"}; the dissolved oxygen files have \code{"DO_"};
the Salinity/Conductivity files have either \code{"Cond_"}, \code{"Con_"} or \verb{"Sal_}
in the name; and that there is only one of each kind of file - four files
total.
}
\examples{
\dontrun{
  paths <- setup_example_dir()
  a <- qc_deployment(paths$deployment)

}

}
