% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_site.R
\name{report_site}
\alias{report_site}
\title{Produce stats and a report for a site and year}
\usage{
report_site(site_dir)
}
\arguments{
\item{site_dir}{Full path to site data (i.e., \verb{<base>/<year>/<site>})}
}
\description{
Produces daily stats and a report with seasonal stats and a number of graphs for a specified site and year.
}
\details{
Two files are written:
\enumerate{
\item \verb{daily_stats_<site>_<year>.csv} - a file with a row for each day of the season with several summary statistics
\item \verb{report_<site>_<year>.pdf} - a PDF report with a table of seasonal stats and a number of graphs
}

\code{stitch_site} must be run first to create the necessary data. If \code{stitch_site} was run at
some point in the past, it is highly advisable to run \code{check_site} to make sure deployment
files haven't changed since the last \code{stitch_site} run.
}
