% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_stats.R
\name{seasonal_stats}
\alias{seasonal_stats}
\title{Produce seasonal stats for a site}
\usage{
seasonal_stats(core)
}
\arguments{
\item{core}{Core data frame, produced by \code{stitch_site}}
}
\value{
List of:
\item{\code{table}}{Data frame with a column of the statistic name and a column with the numeric statistics}
\item{\code{formatted}}{Data frame with a column of the statistic name and a column with the formatted statistics}
}
\description{
Produces a data frame of seasonal stats.
}
\keyword{internal}
