% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch_site.R
\name{stitch_site}
\alias{stitch_site}
\title{Stitch all deployments for a site and year}
\usage{
stitch_site(site_dir, max_gap = 1, report = FALSE)
}
\arguments{
\item{site_dir}{Full path to site data (i.e., \verb{<base>/<year>/<site>}). The path must include QCed results}

\item{max_gap}{Maximum gap to quietly accept between deployments (hours); a msg will be printed if this gap is exceeded}

\item{report}{Run \code{report_site} if TRUE}
}
\description{
Merges all QCed deployments for the specified site and year and writes result files. Missing
dates and times are interpolated, with a warning for gaps that are suspiciously large. Data
that are out range for their sensor are flagged. Writes three versions of the data file for the
complete season, and hash files for use by \code{check_site}.
}
\details{
Three versions of the data file are written to <site_dir>/combined/:
\enumerate{
\item \verb{archive_<site>_<year>.csv} - contains all columns, for complete archival.
\item \verb{WPP_<site>_<year>.csv} - contains only columns required by MassDEP (a.k.a. the "WPP" file).
\item \verb{core_<site>_<year>} - just the good stuff. This is the file used for producing summaries and reports.
}

An additional file is written for internal use by \code{check_site}:
\enumerate{
\item \code{hash.txt} - a tab-delimited file lists paths to deployment files and md5 hashes.
}
}
